Main Menu Template - looks fine, does nothing
Hey guys! So I got the template downloaded and in a scene. I also
change the colors around a bit so it looks nice and gaudy now (YIKES)!
So here’s a bunch of files. We’ll def get rid of some of them like the
audio mixer and stuff, and I’ll go through and pull that stuff out
manually, but here’s something to start with…