IN every man's life there comes a time when he must stand up and check his resources. Today is that day. Today is the day of our resourtic reckoning
Shall we prevail? Will we perish? All will depend on these resources;
if only I hadn’t spent all our hard earned bells on stupid furniture.
If only I hadn’t foolishly sojourned down that dark alley only to be
damaged by a monster and run; if only I had more health. But this is
what the autosave is, so we’ve go to work with the hand we’re dealt.
“I’m always blessed. Bless up.”